Add SF Symbol to denote user's active status
Use the SF Symbol, "person.crop.circle.badge.fill" to denote whether the
user is active or not. If the user is active, style the symbol with
green as the primary color and Color.secondary as the secondary color.
If the user is inactive, display red as the primary color and
Color.secondary as the secondary color.